backend/requirements.txt
langchain_core==0.3.28
langchain_openai==0.2.14
langgraph==0.2.60
python-dotenv==1.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31